Pass "relink" as argument to setGuardedInvocation, don't put it in call site.
"relink" is now being passed as an argument to setGuardedInvocation instead of
being put as a field in the RelinkableCallSite. This allows the call sites to
remain smaller at the expense of recreating the relink method more frequently.

Also, got rid of the callSiteDescriptor argument in relinkAndInvoke() as it
can be extracted from callSite, making the relink method handles slimmer.