Introduced the concept of a "constructor extender".
Constructor extenders allow language runtimes to extend the set of 
constructors that apply when a Class is instantiated using "dyn:new". This is
useful for the implementation of the often recurring idiom where Java 
interfaces and abstract classes can be anonymously implemented by passing a
language runtime native implementation object to a constructor to the Class
representing the (otherwise uninstantiatable) interface or abstract class.