Added "name(type, type, …)" syntax in the method names for callPropWithThis.
This is meant to support the "name(type, type, ...)" syntax that programmers
can use in a method name to manually pin down an exact overloaded variant.
This is not usually required, as the overloaded method resolution works
correctly in almost every situation. However, in presence of many
language-specific conversions with a radically dynamic language, most
overloaded methods will end up being constantly selected at invocation time,
so a programmer knowledgable of the situation might choose to pin down an
exact overload for performance reasons.