Removed the ability to invoke "dyn:new" on java.lang.Class
I now have a sound justification for not allowing "dyn:new" to work on Class
objects. Namely: try to imagine a language that has a concept of classes in
the language, but not in the runtime. I.e. something like C++ without RTTI.
Even in Java, I bet there's a lot of programmers that can write quite
complex programs without ever having to be exposed to the concept of runtime
type information through java.lang.Class and friends. Yet they still use
classes in their source code - i.e. to access static fields and invoke
"new", but they never open the door into the runtime type system with
<typename>.class expression, or Object.getClass(), or Class.forName().

It illustrates that there's a difference between a concept of the class in
the context of the text of the source code of the program (which is what I
ClasStatics is representing), and a runtime reflective object representing a
class (which is java.lang.Class). I'm thinking java.lang.Class as just being
handled as a POJO, and ClassStatics being the only object recognized by
BeansLinker as responding to "dyn:new".

Slightly ironically, since Dynalink concerns itself with dynamic linking, we
need an object that represents the source-code level "class" concept at run
time, but obviously java.lang.Class isn't it (it provides no access to the
statics of the class it represents).