Give priority to language-specific conversions when deciding which type is more specific.
The old logic (giving priority to Java subtype relation) produced unexpected
results when language-specific rules allowed different primitive types, i.e.
when a double can be converted to both double and int, int would've always
be chosen as it's a subtype of double. (Note that with Java rules,
conversion from double to int is not allowed, so in absence of
language-specific conversions, this isn't an issue.)