Move CallSiteDescriptor class from "dynalink.linker" to "dynalink" package.
The policy is that "dynalink" package should declare everything that a user
of the library that is not implementing its own linker might need, and that
the "dynalink.linker" package has additional stuff that implementers of
linkers need.

CallSiteDescriptor is needed by language runtimes that don't provide
linkers, so for API consistency, it needs to live in the "dynalink" package,
not in the "dynalink.linker" package.