Initial support for first-class method support.
This commit introduces the "dyn:getMethod[:name]"(Object)Object operation
that retrieves an object representing a named method on an object. I chose
to separate methods from properties and not return methods from BeanLinker
through "dyn:getProp", hence the new verb that operates only on the method
namespace. For languages that don't discriminate these namespaces, I will
soon be introducing composite operations.

Also new is the "dyn:call" operation that can be applied to any object that
can be called as a method/function. This allows the separation of
"dyn:callPropWithThis" (probably soon to be rethought) into two operations,
"dyn:getMethod" followed by "dyn:call".

Note that "dyn:getMethod" should not bind the returned method to the
receiver, that is, a subsequent "dyn:call" will need to specify it again as
the second argument.

Also, there's no assumption as to what should the object returned by the
"dyn:getMethod" be; the only expectation is that it is possible to
subsequently apply "dyn:call" to it. They don't have to necessarily be
either java.lang.reflect.Method or java.lang.invoke.MethodHandle objects,
they can be other implementation-specific objects as long as there's a
linker in the system that can link a "dyn:call" operation on them.