Add support for composite operations in BeansLinker.
Composite operations are of form "dyn:getElem|getProp|getMethod:foo",
or "dyn:getElem|getProp|getMethod" (with dynamic name). The semantics is to
try to link to a method that will try the operations in the specified order.
Valid combinations are getElem with getProp and getMethod, as well as
setElem with getElem, in any permutation.