Guard against the situation where Dynalink runs with wide set of privileges.
I.e. we could have Dynalink in boot or extension classpath. In this case,
it will have access to classes in restricted packages. If linking to such a
class is first requested by a similarly widely privileged code, we might be
able to introspect and cache restricted methods that we'll later on happily
issue to less privileged code too.

To prevent this, Dynalink will perform SecurityManager.checkPackageAccess()
using a helper class loaded into an isolated secure class loader with no
permissions whatsoever.

As a net effect, Dynalink will not allow any access to code in restricted
packages; classes in restricted packages that have publicly accessible
superclasses and superinterfaces will be accessible through this publicly
visible subset of their API.

Note that this is how Dynalink so far operated when it was running
unprivileged under a security manager; these changes only ensure that it
operates consistently even when it runs with broad privileges under a
security manager.