Fix for DynamicLinker.getRelinkedCallSiteLocation on first call.
DynamicLinker.getRelinkedCallSiteLocation() was incorrectly reporting
the location when the call site was being linked for the first time, as
then the stack is different than when relinking.

Also, the correct name for this API is
DynamicLinker.getLinkedCallSiteLocation(). Introduced the new name and
left the old name in as a deprecated delegating method.