A new approach to applying return value filters.
A new approach to ensuring that return value filters are always applied
when they are needed. We need to apply them just before we link a method
into a call site. We can't let this be the responsibility of individual
linkers (e.g. POJO linker), as they might be unaware of the target (call
site owner) language's type conversion rules. For that reason we:

a) institute a policy where individual linkers don't convert the return
type of the produced method handles to conform to the call site, as
they can't know what's the correct policy for doing it (convert vs.
throw a conversion exception). We only allow them to do it when such
conversion doesn't produce information loss. In practical terms, this
means that method invocation conversions (JLS 5.3) are mostly
allowed, but we even restrict those - we disallow the float > long
and double > long conversions as they lose precision, and we also
disallow any conversion of char type, similar to how conversion of
boolean types is disallowed. Char is similar to boolean in the sense
that it is an enumeration type, not an arithmetic type. While Java
language happily allows arithmetic operations on ints, in dynamic
context there is sufficient ambiguity to make it dangerous, e.g. an
optimistic call site that accepts int return value will inadvertedly
convert 'A' to 65, which in our opinion is two different values.

b) introduce the concept of a pre-link filter, which is really a
f(GuardedInvocation, LinkRequest, LinkerServices)->GuardedInvocation
function. This pre-link filter is configured in the dynamic linker
factory and becomes part of the dynamic linker, to be executed on
every guarded invocation returned from the individual linkers. This
way the language runtime's return value filters will be applied to all
invocations, regardless of what linker they came from.