Allow different fallbacks for switch point invalidations and catches.
It used to be that the fallback method for invalidation of switch points
was also used as the fallback for the catches. Now they can be
different.

ChainedCallSite takes advantage of this, so it will not remove catches
from the link chain when switch points are invalidated, only when an
exception is first thrown.

NOTE: this is a backwards-incompatible change in that it reverses the
meaning of arguments in the GuardedInvocation.compose()!